+++ examples/data-sources/nxos_system/data-source.tf
@@ -1,0 +1,2 @@
+data "nxos_system" "example" {
+}