+++ examples/resources/nxos_system/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_system.example "sys"