+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Add nxos_feature_vpc resource and data source
 - Add nxos_feature_vn_segment resource and data source
 - Add nxos_feature_nv_overlay resource and data source
+- Add nxos_system resource and data source
 
 ## 0.3.0